test(vite): refactor
